Add types table to TagGroup & move notifications & pagination back to /core
