+++ .storybook/NotificationsDecorator.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { NotificationsProvider } from '../src/packages/notifications'
+import { NotificationsProvider } from 'packages/core/Notifications'
 
 const NotificationsDecorator = (Story: any) => (
   <NotificationsProvider>
