+++ src/stories/core/Pagination.stories.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 
 // Components
-import Pagination, { PaginationProps } from "packages/pagination/Pagination";
+import Pagination, { PaginationProps } from "packages/core/Pagination/Pagination";
 
 export const Default = (props: PaginationProps) => {
   return <Pagination max={4} total={256} onChange={console.log} {...props} />;
